Fix error on 'validationMessage' custom binding.
When the 'validationMessage' custom binding is used in an element, and
that element is later removed from the DOM, if the validation is
triggered an error is thrown.

This PR fixes that problem by checking on whether the element has a
parent node before updating it, and by disposing the subscription that
updates the element when the element is disposed.